requirements.txt
Django==1.6.5
Pillow==2.4.0
South==0.8.4

#psycopg2==2.5.2

python-dateutil==2.1

django-compressor==1.3
django-cors-headers==0.10
django-extensions==1.3.3
django-filebrowser
